requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
python-status==1.0.1
requests==2.18.4
urllib3==1.22
